(D01) Got part 1 working with the example.
I'm not that happy with the file parsing code, but I'm quite happy
with the Dial struct implementation (including the associated
functions).

The example works fine, and I'm alerting when there is bad input.

I could consider improving the parsing code when I get more
adept at and familiar with Rust.