(D01) I completed Part 2 with great success.
But I still feel like the left() and right() functions are a little
kludgey. Perhaps in part because I chose to use u32 to track position
and clicks, rather than i32.

It could use some more attention. But we have to move forward.
Summarize the change in less than 50 characters

Because:
- Explain the reasons you made this change
- Make a new bullet for each reason
- Each line should be under 72 characters

Explain exactly what was done in this commit with more depth than the
50 character subject line. Remember to wrap at 72 characters!

Include any additional notes, relevant links, or co-authors.