(D02) successful completion
I implemented an invalid_ids() method that returned a vector of the
invalid IDs for an IdRange object. I think it is about as fast as it
could be, but feels kludgey. Note that I had to accumulate the
sum of product IDs into an u64 as a u32 wouldn't hold the value.