(D02) Implemented Part 2, but with a bug
I implemented Part 2, with a nice bit of code that almost works
correctly. The remaining problem is that the product ID 222222 is
matched by a pattern of repeated 2s, 22s, and 222s ... so it shows
up in the output three times. I'm thinking about solving this with
a hash collection or similar uniquely keyed collection.