(D04) still working on the grid struct impl
I'm stuck on producing an Error when an invalid
character is read from the input. Not solved,
but I need to take a break.
Summarize the change in less than 50 characters

Because:
- Explain the reasons you made this change
- Make a new bullet for each reason
- Each line should be under 72 characters

Explain exactly what was done in this commit with more depth than the
50 character subject line. Remember to wrap at 72 characters!

Include any additional notes, relevant links, or co-authors.