(D04) Added the code to populate the grid
Added all the code to populate the grid from the input file.

Next up is to calculate the number of neighbors of any cell.
Summarize the change in less than 50 characters

Because:
- Explain the reasons you made this change
- Make a new bullet for each reason
- Each line should be under 72 characters

Explain exactly what was done in this commit with more depth than the
50 character subject line. Remember to wrap at 72 characters!

Include any additional notes, relevant links, or co-authors.