(D04) Completed Part 2, successfully.
I updated the specs with the part 2 information, and implemented
a solution (which required adding grid modification methods).

I determined the correct answer for the example, and also for the
regular input file. The input file uses a 140x140 grid, so the
iteration through removing rolls took many iterations. I could
probably make the code more efficient, but I submitted a correct
Part 2 solution, so that is all for today.