(D05) I think I've implemented part 2
I also made Part 1 more efficient. The secret
was merging the overlapping ranges. It would be
even faster now if I'd sort and index the ranges by ascending start values and end values, but I'm
not going to take the time to do that.
Summarize the change in less than 50 characters

Because:
- Explain the reasons you made this change
- Make a new bullet for each reason
- Each line should be under 72 characters

Explain exactly what was done in this commit with more depth than the
50 character subject line. Remember to wrap at 72 characters!

Include any additional notes, relevant links, or co-authors.