(D08) Getting closer, but not yet working
Still trying to get Part 1 solved. I'm close, but not getting the
right answer. There is one pairing I seem to be missing, ignoring,
skipping. With the example data I get circuits sized 4,3,2,2,2 and
it should be 5,4,2,2.
Summarize the change in less than 50 characters

Because:
- Explain the reasons you made this change
- Make a new bullet for each reason
- Each line should be under 72 characters

Explain exactly what was done in this commit with more depth than the
50 character subject line. Remember to wrap at 72 characters!

Include any additional notes, relevant links, or co-authors.