(D08) Removed sqrt operation from distance calc
I removed the square root operation from the distance calculation.
We are only comparing distances, so the sum of squares value
is sufficient and we don't need the actual distance value.

This allows removing all f64 types and replacing them with i64
and u64 types.

Unfortunately, the bug remains. Still getting 4,3,2,2,2 circuits
from the example input; should be 5,4,2,2.