(D09) Finished part 1
Just a basic recursive search of all red tile pairs.

This time I did the unit test route before implementing the CLI.
And I put the file and string parsing code in distinct functions,
which cleans up the unit test code and CLI main function.

This was the fastest I completed a part 1, to date.